requirements.txt
humanize=>4.9.0
py-dateutil>=2.2
python-dotenv>=1.0.1
requests=>2.31.0
rich=>13.7.0
tgtg>=0.16.1